// src/app/api/model/watchProviderItem.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface WatchProviderItem { 
    logo_path?: string;
    provider_id?: number;
    provider_name?: string;
    display_priority?: number;
}