// src/app/api/model/watchProviderList.ts
/**
 * TMDB API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WatchProviderCountry } from './watchProviderCountry';


export interface WatchProviderList { 
    id?: number;
    results?: { [key: string]: WatchProviderCountry; };
}